+++ src/containerapp/HISTORY.rst
@@ -5,6 +5,9 @@ Release History
 
 0.3.11
 ++++++
+* 'az containerapp create': Add keda scale rule parameters
+* 'az containerapp update': Add keda scale rule parameters
+* 'az containerapp revision copy': Add keda scale rule parameters
 
 
 0.3.10